move Thresholds to Test Options popover
